+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 1.0.0-beta.2
 
+* Add a `pkg-standalone-dev` task for building a script that can be invoked for
+  testing.
+
 * Fix a bug where the version variable wouldn't be set for certain executables.
 
 # 1.0.0-beta.1